introducing Grouping/Setlists
+ make Groups as multivalue (separate table)
+ integrate grouping functionality in Retail Screen
+ Import fails randomly...workaround Import again (don't clear the DB)
+ Tool is based on Rocksmith Toolkit 16th Oct version
+ Edit of Arragements and Tones DB can only be done from Access for the
moment
+ add date as duplication text
+ small improv: Any change of song in the list should also save :)
+ cleanup/hidden the interface even more
+ cleanup double saving issue

+++ RocksmithToolkitLib/DLCPackage/DLCPackageData.cs
@@ -636,7 +636,7 @@ namespace RocksmithToolkitLib.DLCPackage
             if (!String.IsNullOrEmpty(audioPreviewPath))
             {
                 var newPreviewFileName = Path.Combine(Path.GetDirectoryName(audioPath), String.Format("{0}_preview{1}", Path.GetFileNameWithoutExtension(audioPath), Path.GetExtension(audioPath)));
-                File.Move(audioPreviewPath, newPreviewFileName);
+                if (!File.Exists(newPreviewFileName)) File.Move(audioPreviewPath, newPreviewFileName); //bcapi as some original create an error here
                 data.OggPreviewPath = newPreviewFileName;
             }
